Compiler - services configuration and dotvvm startup resolving changed
