+++ src/DotVVM.Compiler/Properties/launchSettings.json
@@ -6,7 +6,7 @@
     },
     "NetCoreApp": {
       "commandName": "Project",
-      "commandLineArgs": "--debugger --json \"{\\\"WebSiteAssembly\\\":\\\"C:\\\\dev\\\\dotvvm-web\\\\src\\\\DotvvmWeb\\\\DotvvmWeb.Docs.AspNetCore\\\\bin\\\\debug\\\\netcoreapp2.0\\\\DotvvmWeb.Docs.AspNetCore.dll\\\",\\\"WebSitePath\\\":\\\"C:\\\\dev\\\\dotvvm-web\\\\src\\\\DotvvmWeb\\\\DotvvmWeb.Docs.AspNetCore\\\",\\\"FullCompile\\\":false,\\\"SerializeConfig\\\":true,\\\"DothtmlFiles\\\":null ,\\\"OutputResolvedDothtmlMap\\\":true,\\\"CheckBindingErrors\\\":true, OutputPath: \\\"C:\\\\temp\\\\dotvvm-compiler\\\"}\""
+      "commandLineArgs": "--debugger --json \"{\\\"WebSiteAssembly\\\":\\\"C:\\\\dev\\\\dotvvm\\\\src\\\\DotVVM.Samples.BasicSamples.AspNetCore\\\\bin\\\\Debug\\\\netcoreapp2.0\\\\DotVVM.Samples.BasicSamples.AspNetCore.dll\\\",\\\"WebSitePath\\\":\\\"C:\\\\dev\\\\dotvvm\\\\src\\\\DotVVM.Samples.BasicSamples.AspNetCore\\\",\\\"FullCompile\\\":false,\\\"SerializeConfig\\\":true,\\\"DothtmlFiles\\\":[],\\\"OutputResolvedDothtmlMap\\\":true,\\\"CheckBindingErrors\\\":true}\""
     }
   }
 }
