+++ src/DotVVM.Compiler/Resolving/NoServicesStartupClassExecutor.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace DotVVM.Compiler.Resolving
+{
+    public class NoServicesStartupClassExecutor : IServicesStartupClassExecutor
+    {
+        public void ConfigureServices(IServiceCollection collection)
+        {
+
+        }
+    }
+}
